NTP configuration on install and fixed tmux attach problems
Why:

We had the wrong timezone and the tmux attach didn't work correctly.
We also had the irssi folder hidden in the config folder.

This change addresses the need by:

Added some script lines in install.sh to fix the problems with ntp
and changed the line that attach to tmux session.