Added build Vim for Mac OS X
Why:

With the new macbook we needed vim with lua capabilities in order to
be able to run some plugins.

This change addresses the need by:

Adding OS detection to perform different installation procedures
depending on the system. The script performs the build process
inside the /temp folder. The installation is performed into
/opt/local for security in Mac OS (Nothing should be removed from
/usr/local) to avoid dependency problems. Once the build is
finished some links will be placed into /usr/local or the $PATH
will be set with preference to /opt/local.

NOTE: The code is not yet 100% functional.