Added Python support to vim build
Why:

Some pluguins like ultisnips or the javascript linter need python
in order to work.

This change addresses the need by:

We added a variable into the build script that changes depending on
the system we are in. As a future feature we could add parameters
to specify this paths.

We also added minor ergonomic fixes to our vim configuration,
specified vimdiff as the default merge and diff tool for git and
solved a bug with ls alias into bashrc.